fix(veil): #2515: use infinite query in candlesticks request
