+++ apps/veil/src/pages/trade/ui/chart/loading-chart.tsx
@@ -1,15 +1,4 @@
-import cn from 'clsx';
-import { useEffect, useRef, useState } from 'react';
-import { createChart, IChartApi } from 'lightweight-charts';
-import { theme } from '@penumbra-zone/ui/theme';
-import { Text } from '@penumbra-zone/ui/Text';
-import { useCandles } from '../api/candles';
-import { observer } from 'mobx-react-lite';
-import { DurationWindow, durationWindows } from '@/shared/utils/duration.ts';
-import { CandleWithVolume } from '@/shared/api/server/candles/utils';
-import { BlockchainError } from '@/shared/ui/blockchain-error';
-
-const ChartLoadingState = () => {
+export const ChartLoadingState = () => {
   return (
     <>
       <style>
@@ -117,143 +106,3 @@ const ChartLoadingState = () => {
     </>
   );
 };
-
-const ChartData = observer(({ candles }: { candles: CandleWithVolume[] }) => {
-  const chartElRef = useRef<HTMLDivElement>(null);
-  const chartRef = useRef<IChartApi>(undefined);
-  const seriesRef = useRef<ReturnType<IChartApi['addCandlestickSeries']>>(undefined);
-  const volumeSeriesRef = useRef<ReturnType<IChartApi['addHistogramSeries']>>(undefined);
-
-  // Initialize the chart once when the component mounts
-  useEffect(() => {
-    if (chartElRef.current && !chartRef.current) {
-      chartRef.current = createChart(chartElRef.current, {
-        autoSize: true,
-        layout: {
-          textColor: theme.color.text.primary,
-          background: {
-            color: 'transparent',
-          },
-        },
-        grid: {
-          vertLines: {
-            color: theme.color.other.tonalStroke,
-          },
-          horzLines: {
-            color: theme.color.other.tonalStroke,
-          },
-        },
-        timeScale: {
-          timeVisible: true,
-          secondsVisible: false,
-          uniformDistribution: true,
-        },
-      });
-
-      // Initialize the candlestick series
-      seriesRef.current = chartRef.current.addCandlestickSeries({
-        upColor: theme.color.success.light,
-        downColor: theme.color.destructive.light,
-        borderVisible: false,
-        wickUpColor: theme.color.success.light,
-        wickDownColor: theme.color.destructive.light,
-      });
-
-      // Set the price scale margins for the candlestick series
-      seriesRef.current.priceScale().applyOptions({
-        autoScale: true,
-      });
-
-      // Initialize the volume series
-      volumeSeriesRef.current = chartRef.current.addHistogramSeries({
-        color: theme.color.success.light + '80',
-        priceFormat: {
-          type: 'volume',
-        },
-        priceScaleId: '',
-        lastValueVisible: false,
-        priceLineVisible: false,
-      });
-
-      // Set the price scale margins for the candlestick series
-      volumeSeriesRef.current.priceScale().applyOptions({
-        scaleMargins: {
-          top: 0.8, // highest point of the series will be 70% away from the top
-          bottom: 0,
-        },
-      });
-
-      // Update volume colors based on price movement
-      volumeSeriesRef.current.setData(
-        candles.map(candle => ({
-          time: candle.ohlc.time,
-          value: candle.volume,
-          color:
-            candle.ohlc.close >= candle.ohlc.open
-              ? theme.color.success.light + '80'
-              : theme.color.destructive.light + '80',
-        })),
-      );
-    }
-
-    return () => {
-      if (chartRef.current) {
-        chartRef.current.remove();
-        chartRef.current = undefined;
-      }
-    };
-  }, [chartElRef, candles]);
-
-  // Update chart when candles change
-  useEffect(() => {
-    if (seriesRef.current && volumeSeriesRef.current) {
-      // Set OHLC data
-      seriesRef.current.setData(candles.map(c => c.ohlc));
-
-      // Set volume data with colors based on price movement
-      volumeSeriesRef.current.setData(
-        candles.map(candle => ({
-          time: candle.ohlc.time,
-          value: candle.volume,
-          color:
-            candle.ohlc.close >= candle.ohlc.open
-              ? theme.color.success.light + '80'
-              : theme.color.destructive.light + '80',
-        })),
-      );
-    }
-  }, [candles]);
-
-  return <div className='h-full w-full' ref={chartElRef} />;
-});
-
-export const Chart = observer(() => {
-  const [duration, setDuration] = useState<DurationWindow>('1d');
-  const { data, isLoading, error } = useCandles(duration);
-
-  return (
-    <div className='flex h-full min-h-0 flex-col'>
-      <div className='flex border-b border-b-other-solid-stroke px-3'>
-        {durationWindows.map(w => (
-          <button
-            key={w}
-            type='button'
-            className={cn(
-              'flex items-center rounded px-1.5 py-3 transition-colors hover:bg-action-hover-overlay hover:text-text-primary',
-              w === duration ? 'bg-action-active-overlay text-text-primary' : 'text-text-secondary',
-            )}
-            onClick={() => setDuration(w)}
-          >
-            <Text detail>{w}</Text>
-          </button>
-        ))}
-      </div>
-
-      <div className='flex min-h-0 grow items-center justify-center'>
-        {error && <BlockchainError direction='column' />}
-        {!error && isLoading && <ChartLoadingState />}
-        {!error && !isLoading && data && <ChartData candles={data} />}
-      </div>
-    </div>
-  );
-});
